Move poly creation to OsmGeom
