+++ tools/OsmGeom.py
@@ -20,6 +20,8 @@
 ##                                                                       ##
 ###########################################################################
 
+import re
+
 # Read a polygon from file
 # NB: holes aren't supported yet
 def read_polygon_wkt(f):
@@ -68,3 +70,48 @@ def read_multipolygon_wkt(f):
     wkt = "MULTIPOLYGON (" + ",".join(polygons) + ")"
 
     return wkt
+
+
+def write_polygon(f, wkt, p):
+
+    match = re.search(r"^\(\((?P<pdata>.*)\)\)$", wkt)
+    pdata = match.group("pdata")
+    rings = re.split(r"\),\(", pdata)
+
+    first_ring = True
+    for ring in rings:
+        coords = re.split(",", ring)
+
+        p = p + 1
+        if first_ring:
+            f.write(str(p) + "\n")
+            first_ring = False
+        else:
+            f.write("!" + str(p) + "\n")
+
+        for coord in coords:
+            ords = coord.split()
+            f.write("  %-11s  %s\n" % (ords[0], ords[1]))
+
+        f.write("END\n")
+
+    return p
+
+def write_multipolygon(f, wkt):
+
+    match = re.search(r"^MULTIPOLYGON\((?P<mpdata>.*)\)$", wkt)
+
+    if match:
+        mpdata = match.group("mpdata")
+        polygons = re.split(r"(?<=\)\)),(?=\(\()", mpdata)
+
+        p = 0
+        for polygon in polygons:
+            p = write_polygon(f, polygon, p)
+
+        return
+
+    match = re.search("^POLYGON(?P<pdata>.*)$", wkt)
+    if match:
+        pdata = match.group("pdata")
+        write_polygon(f, pdata, 0)
